Updates the game to allow of .NET 4.0 AI assmeblies
Previously I had to downgrade the game to target .NET 3.5 since when I first created it 4.0 was only just released (yeah everyone wasn't as fast as I was to install it). But now it was time to get back up in versions. Although I even downgraded BattleShip.Interface to target .NET 2.0 to allow for building AI assemblies with lower versions of the framework. Not very useful perhaps, but it might in the future.. for some reason... maybe?